pyproject.toml
[project]
name = "sql-raq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.134.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]
